Configurations:  'specification/redisenterprise/resource-manager/Microsoft.Cache/RedisEnterprise/tspconfig.yaml', API Version: 2026-02-01-preview, SDK Release Type: beta, and CommitSHA: 'c60ab03fa5dae61d40274ea4943caeb0dc02c0cd' in SpecRepo: 'https://github.com/Azure/azure-rest-api-specs' Pipeline run: https://dev.azure.com/azure-sdk/internal/_build/results?buildId=6426772 Refer to https://eng.ms/docs/products/azure-developer-experience/develop/sdk-release/sdk-release-prerequisites to prepare for SDK release.

+++ sdk/redisenterprise/azure-resourcemanager-redisenterprise/src/main/java/com/azure/resourcemanager/redisenterprise/models/SkuDetails.java
@@ -24,7 +24,7 @@ public final class SkuDetails implements JsonSerializable<SkuDetails> {
     /*
      * The cache size in GB.
      */
-    private Double sizeInGB;
+    private Float sizeInGB;
 
     /**
      * Creates an instance of SkuDetails class.
@@ -46,7 +46,7 @@ public final class SkuDetails implements JsonSerializable<SkuDetails> {
      * 
      * @return the sizeInGB value.
      */
-    public Double sizeInGB() {
+    public Float sizeInGB() {
         return this.sizeInGB;
     }
 
@@ -77,7 +77,7 @@ public final class SkuDetails implements JsonSerializable<SkuDetails> {
                 if ("name".equals(fieldName)) {
                     deserializedSkuDetails.name = reader.getString();
                 } else if ("sizeInGB".equals(fieldName)) {
-                    deserializedSkuDetails.sizeInGB = reader.getNullable(JsonReader::getDouble);
+                    deserializedSkuDetails.sizeInGB = reader.getNullable(JsonReader::getFloat);
                 } else {
                     reader.skipChildren();
                 }
